requirements.txt
customtkinter
openai>=1.50.0
python-dotenv
PyPDF2